# docs/assets/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReativacaoIn": {
        "properties": {
          "aceita": {
            "title": "Aceita",
            "type": "boolean"
          }
        },
        "required": [
          "aceita"
        ],
        "title": "ReativacaoIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "api__routers__leads__CorretorIn": {
        "properties": {
          "corretor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corretor Id"
          }
        },
        "title": "CorretorIn",
        "type": "object"
      }
    },
    "securitySchemes": {
      "Token do painel": {
        "description": "Perfil local: o valor de `SDR_PAINEL_TOKEN` (em desenvolvimento, `dev-token`). Perfil AWS: o JWT emitido pelo Cognito.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Projeto Mora — Vértice Imóveis (POC acadêmica)"
    },
    "description": "\nAPI do **Mora**, o agente SDR imobiliário da Vértice Imóveis.\n\nA etiqueta de cada rota diz quem a usa e o que ela exige:\n\n| Etiqueta | Quem usa | Autenticação |\n|---|---|---|\n| `público` | site vitrine e chat | nenhuma |\n| `corretor` | painel do corretor | `Authorization: Bearer …` |\n| `operacao` | rotinas de atendimento no painel | `Authorization: Bearer …` |\n| `admin` | configuração, governança de IA e auditoria | `Authorization: Bearer …` |\n\n**Como autenticar no perfil local:** clique em **Authorize** e informe o valor de\n`SDR_PAINEL_TOKEN` (no ambiente de desenvolvimento, `dev-token`). No perfil AWS o token é o\nJWT do Cognito, validado pelo authorizer do API Gateway antes de chegar aqui.\n\nToda alteração feita por estas rotas é registrada em auditoria (ver `GET /auditoria`).\n",
    "license": {
      "name": "Uso acadêmico — licença a confirmar"
    },
    "title": "Mora — SDR Imobiliário API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/interesses/lead/{lead_id}": {
      "get": {
        "description": "Imóveis que passaram pela conversa deste lead, do mais recente para o mais antigo.",
        "operationId": "do_lead_interesses_lead__lead_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "lead_id",
            "required": true,
            "schema": {
              "title": "Lead Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Do Lead",
        "tags": [
          "corretor"
        ]
      }
    },
...
    "/leads": {
      "get": {
        "operationId": "listar_leads_get",
        "parameters": [
          {
            "in": "query",
            "name": "estagio",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estagio"
            }
          },
          {
            "in": "query",
            "name": "temperatura",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Temperatura"
            }
          },
          {
            "in": "query",
            "name": "corretor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Corretor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Listar",
        "tags": [
          "corretor"
        ]
      }
    },
    "/leads/crm/sync": {
      "post": {
        "description": "Simula integração com CRM externo (diferencial): exporta leads qualificados/agendados.",
        "operationId": "crm_sync_leads_crm_sync_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Crm Sync",
        "tags": [
          "corretor"
        ]
      }
    },
    "/leads/{lead_id}": {
      "get": {
        "operationId": "detalhe_leads__lead_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "lead_id",
            "required": true,
            "schema": {
              "title": "Lead Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Detalhe",
        "tags": [
          "corretor"
        ]
      }
    },
    "/leads/{lead_id}/analisar": {
      "post": {
        "description": "Pede ao agente (worker Resumidor) um novo briefing + análise de sentimento/perfil. Assíncrono: o painel recarrega.",
        "operationId": "analisar_leads__lead_id__analisar_post",
        "parameters": [
          {
            "in": "path",
            "name": "lead_id",
            "required": true,
            "schema": {
              "title": "Lead Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Analisar",
        "tags": [
          "corretor"
        ]
      }
    },
    "/leads/{lead_id}/corretor": {
      "put": {
        "description": "Atribui (ou troca/remove) o corretor responsável pelo lead; visitas futuras acompanham.",
        "operationId": "atribuir_leads__lead_id__corretor_put",
        "parameters": [
          {
            "in": "path",
            "name": "lead_id",
            "required": true,
            "schema": {
              "title": "Lead Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api__routers__leads__CorretorIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Atribuir",
        "tags": [
          "corretor"
        ]
      }
    },
    "/leads/{lead_id}/mensagens": {
      "get": {
        "operationId": "mensagens_leads__lead_id__mensagens_get",
        "parameters": [
          {
            "in": "path",
            "name": "lead_id",
            "required": true,
            "schema": {
              "title": "Lead Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Mensagens",
        "tags": [
          "corretor"
        ]
      }
    },
    "/leads/{lead_id}/reativacao": {
      "put": {
        "description": "Liga/desliga o aviso de imóvel novo para este lead.\n\nO cliente pode desligar sozinho pela conversa (\"não quero mais avisos\"), mas o pedido também\nchega por telefone, por e-mail e no meio de uma visita — e aí quem registra é o corretor. Sem\nesta rota, honrar um opt-out exigiria UPDATE no banco, que é o mesmo que não honrar.",
        "operationId": "preferencia_de_reativacao_leads__lead_id__reativacao_put",
        "parameters": [
          {
            "in": "path",
            "name": "lead_id",
            "required": true,
            "schema": {
              "title": "Lead Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReativacaoIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Preferencia De Reativacao",
        "tags": [
          "corretor"
        ]
      }
    },
    "/notificacoes": {
      "get": {
        "operationId": "listar_notificacoes_get",
        "parameters": [
          {
            "in": "query",
            "name": "apenas_nao_lidas",
            "required": false,
            "schema": {
              "default": false,
              "title": "Apenas Nao Lidas",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "title": "Limite",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Listar",
        "tags": [
          "operacao"
        ]
      }
    },
    "/notificacoes/lidas": {
      "post": {
        "operationId": "marcar_todas_notificacoes_lidas_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Marcar Todas",
        "tags": [
          "operacao"
        ]
      }
    },
    "/notificacoes/{notificacao_id}/lida": {
      "post": {
        "operationId": "marcar_lida_notificacoes__notificacao_id__lida_post",
        "parameters": [
          {
            "in": "path",
            "name": "notificacao_id",
            "required": true,
            "schema": {
              "title": "Notificacao Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Marcar Lida",
        "tags": [
          "operacao"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Perfil local (docker compose)",
      "url": "http://localhost:8000"
    },
    {
      "description": "Perfil AWS — a confirmar no deploy",
      "url": "https://api.exemplo.com.br"
    }
  ],
...
}